Convert Dependency Graph to  instance variable instead of static
We need to create multiple instances of GraphResolver class based on whether we want to resolve Python or C dependencies.
For each of those instances we need a separate dependency graph (e.g. one for python and one for C). This can be achieved by making the corresponding variables instances instead of static.